requirements.txt
tensorflow
keras
pyyaml
numpy
matplotlib
Pillow
Pandas
nltk